Anh hao

+++ QuanLyChuoiCuaHangTienLoi_Ministop/ET/ET_LichSuThayDoiGia.cs
@@ -11,8 +11,8 @@ namespace ET
         public string MaLichSu { get; set; }
         public string MaSanPham { get; set; }
         public string MaCuaHang { get; set; }
-        public double GiaCu { get; set; }
-        public double GiaMoi { get; set; }
+        public decimal GiaCu { get; set; }
+        public decimal GiaMoi { get; set; }
         public string NguoiThayDoi { get; set; }
         public string LyDo {  get; set; }
         public DateTime NgayThayDoi { get; set; }
